attach to use /tmp directory ; file structure changes
